git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@8820 01821946-851b-0410-b7f4-f1477c70e57a

+++ gameio/loadtextures.cpp
@@ -731,7 +731,7 @@ return FindHiresBitmap (szFile, nFile, bD1) != NULL;
 static bool HaveHiresModel (const char* bmName)
 {
 if (gameStates.app.bStandalone) {
-	for (int i = 0; i < sizeofa (szPowerupTextures); i++)
+	for (int i = 0; i < (int) sizeofa (szPowerupTextures); i++)
 		if (strstr (bmName, szPowerupTextures [i]))
 			return true;
 	}
